Fixed now playing overlay overlapping 5-day weather in Standard Horizontal BasicHome.
Updated Horizontal Home writer to use different common.overlay.basichome.xml depending if using Context Labels. When context labels are used the Now Playing overlay is offset -90 so it appears above the menu rather on it.

Note: These changes assume the y-pos is not adjusted by the user...probably best to put some smarts in for that (Trevor?)