Update to Music NoPlaying screens
Changed the control used to display Personal Rating back to list control. 
Added additional define in the main files to control the scaling of the above control; The setting <define>#star.rating.size:32</define> in the MyMusicNowPlaying.xml and the AU versions can be set to 32 for 1280x720 or 48 for 1920x1080 displays - default setting is 32 for 1280x720.
Fixed some control visibility issues for Fanart Handler 
Updated progress bar code to support latest MP SVN 
Changed the full screen Menu layout