Fixed not being able to select videowindow for nowplaying videos in Vertical Basichome.
When navigating right from Shutdown button, id 17 (videowindow) will be selectable if video media is playing.

Please review. I could not make the audio selectable button and video button selectable as I don't know how to navigate to one or the other depending on which is playing so I chose the video one.