Fixed nlog not being reference counted during install, was being uninstalled even though was previously installed.
Setup detects if a file is already installed before installing, if already installed a reference count is incremented and is checked on uninstall to see if file should be removed. In the case user had movingpictures installed prior to installing streamedmp then on uninstall of streamedmp it would remove nlog but leave movingpictures.dll.