Testable version of the StreamedMPConfig plugin. It works by reading the SkinUpdate.xml directly from the SVN tree ("http://streamedmp.googlecode.com/svn/trunk/SkinUpdate/SkinUpdate.xml"), this file contains the version of the update and links to the change log and archive files.
To aid with testing it will 1st look at the root of C: for the SkinUpdate.xml and ChangeLog.rtf files and use these if found, allowing for testing without making a mess of svn.

There is some more polishing to do, tidy up the update GUI screen, more infomation, confrm dialog etc.