Fixed cache not being cleared on unattended patch installs.
Handle UnauthorisedAccessException when trying to copy files to program files directory, most likely because of UAC. 

I did a fresh install of MediaPortal and UAC enabled on Vista and had issues. Maybe its best to add the manifest to prompt elevation or just leave it as is with handled exception?