test program - this will calculate and add a checksum to all xml files in the live StreamedMP directory, in addition it reformats the files as it goes (saving 300k).
I see this as the next stage with the patching but will  required a major release before it can be realised. If we checksum all xml files we have a method of checking if users have modified files and if they have we can not update these and provide a copy of the patched files elsewhere.

The main part is a class that can be added to any program to add, read and compare the checksums on files.