Issue 346: Part I, installer now installs all TVSeries and MovingPictures Mods. Installer and (future GUI) now just modify the <imports> in the TVSeries.xml and MovingPictures.xml to point to the correct style/mod.
The installer/GUI plugin can check the top two defines in each XML to determine what is currently being used e.g.

TVSeries:
<define>#defaultstyle:yes</define>
#defaultstyle=yes when default style or no when fanart style

<define>#widebannerstyle:0</define>
0: Default (depends on Default or Fanart Style)
1: 5x2 Grid + Poster
2: 5x3 Grid
3: 7x3 Grid
4: 10x4 Grid

MovingPictures:
<define>#defaultstyle:Yes</define>
#defaultstyle=yes when default style or no when fanart style

<define>#thumbnailstyle:2</define>
0: Default (depends on Default or Fanart Style)
1: SmallIcon [12x4], Large Icon [8x3] Grid
2: SmallIcon [12x3], Large Icon [8x2] Grid + Info

When the user has selected a Style and a Thumbnail Mod the Installer/GUI plugin sets the <defines> and modifies the <imports>. Each import that needs to be modified has a tag so it can be identified and set using XPATH syntax e.g.

<import tag="Facade">TVSeries.Default.Facade.xml</import> 

can be referenced using this path in code: "/window/controls/import[@tag='Facade']"