Reworked logic for music overlays on basichome, home and common overlays.
Removed logic which checks if the current thumb is empty or not as there was no consistency with how it works in different music views and screens! The current logic is now much simplier and will fall through the images e.g.

You will see the default logo if no artist and album images.
You will see the artist image if there is an artist image but no album image.
You will see an animation of artist and album image when both exist.

Note: only problem with this logic is that you see extra images when fading in / out of windows during window open/close animations.